route and fuzzy search implimented
moved the fuzzy_search folder to be within the backend folder.
created the flask route function to handle the frontend fetch request
connected the fuzzy search module to the backend basic search - only used when specified within fetch request
added additional movies to the movie database
